backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.32,<0.33
sqlalchemy>=2.0,<2.1
alembic>=1.14,<1.15
psycopg2-binary>=2.9,<2.10
python-jose[cryptography]>=3.3,<3.4
bcrypt>=4.0,<5.0
python-multipart>=0.0.12,<0.0.13
boto3>=1.35,<1.36
pillow>=10.4,<10.5
pydantic-settings>=2.5,<2.6
email-validator>=2.1,<2.2
httpx>=0.27,<0.28
